refactor: complete move of activation to process
After proceeding piecemeal it ended up that the whole activation subdir was actually generic except for the naming of entities.
