+++ tests/integration/services/activation/test_manager.py
@@ -23,15 +23,15 @@ from pytest_lazyfixture import lazy_fixture
 
 from aap_eda.core import models
 from aap_eda.core.enums import ActivationStatus
-from aap_eda.services.activation.manager import (
+from aap_eda.services.process.engine import exceptions as engine_exceptions
+from aap_eda.services.process.engine.common import ContainerEngine
+from aap_eda.services.process.manager import (
     ACTIVATION_PATH,
     LOGGER,
     ActivationManager,
     AnsibleRulebookCmdLine,
     exceptions,
 )
-from aap_eda.services.process.engine import exceptions as engine_exceptions
-from aap_eda.services.process.engine.common import ContainerEngine
 
 
 def apply_settings(settings: SettingsWrapper, **kwargs):
